Move parseWsUrl and createWsUrl to a new package
The 'protocol' package imports txaio, and so unneccesarily
requires anyone who uses "parseWsUrl" (e.g. a test) to initialize
txaio first (and thereby choose twisted vs asyncio).

There may be other candidates in 'protocol' to move. Note that
of course this breaks any existing imports of these functions.
